refactor(sync): split post-merge-validator into per-validator modules (sync-1vf5)
Tier B ratchet: drop packages/sync/src LOC cap from 1100 to 750. Split
post-merge-validator.ts (1096 LOC) by moving each validator into its own
file under validators/:

- internal.ts — DocRecord/ArchivableEntity/ParentableEntity/SortableEntity
  types, getEntityMap, getParentId, ENTITY_FIELD_MAP, getEntityTypeByFieldName
- tombstones.ts — enforceTombstones
- hierarchy-cycles.ts — detectHierarchyCycles + detectCyclesForField
- sort-order.ts — normalizeSortOrder + collectSortOrderPatches +
  partitionByGroupField
- check-in.ts — normalizeCheckInRecord
- friend-connection.ts — normalizeFriendConnection
- fronting.ts — normalizeFrontingSessions + normalizeFrontingCommentAuthors
- timer-config.ts — normalizeTimerConfig
- webhook-config.ts — normalizeWebhookConfigs + VALID_WEBHOOK_EVENT_TYPES
- bucket-content-tags.ts — validateBucketContentTags

post-merge-validator.ts (270) keeps runAllValidations + re-exports every
validator so existing test imports (../post-merge-validator.js) and the
package index keep resolving unchanged. 951 sync tests pass.
